chore(Airbnb): rename client helper file
